requirements.txt
Faker
pandas
SQLAlchemy
psycopg2-binary
python-dotenv